实现fp16推理

+++ utils/functions.py
@@ -83,10 +83,11 @@ def transform(image: np.ndarray, openvino_preprocess: bool = False) -> np.ndarra
     Returns:
         np.ndarray: 经过预处理的图片
     """
-    image = image.transpose(2, 0, 1).astype(np.float32)        # [H, W, C] -> [C, H, W]
+    image = image.transpose(2, 0, 1)        # [H, W, C] -> [C, H, W]
 
     # openvino预处理会自动处理scale
     if not openvino_preprocess:
+        image = image.astype(np.float32)
         image /= 255.0                      # 归一化
 
     return np.expand_dims(image, 0)  # [C, H, W] -> [B, C, H, W]
